indicator_engine/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
psycopg2-binary==2.9.10
numpy==1.26.4
pandas==2.2.2
ta-lib==0.4.32
requests==2.32.3
apscheduler==3.10.4
